requirements.txt
ccxt==1.58.96
pandas==1.3.5
ta==0.9.0
python-crontab==2.6.0
